requirements.txt
appdirs==1.4.0
Django==1.10.5
django-appconf==1.0.2
django-imagekit==3.3
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
